now details fetch from db
